n85 rocket.io

+++ app/src/main/java/com/viatom/checkme/bean/TmpBean.kt
@@ -6,6 +6,6 @@ data class TmpBean(
     var date: Date = Date(),
     var timeString: String = "",
     var way: Int = 0,
-    var tmp: Int = 0,
+    var tmp: Float = 0f,
     var face: Int = 0
 )
